# Benchmark Report for *YaoBlocks*

## Job Properties
* Time of benchmark: 18 May 2019 - 17:30
* Package commit: 17b00b
* Julia commit: 80516c
* Julia command flags: None
* Environment variables: None

## Results
Below is a table of this job's results, obtained by running the benchmarks.
The values listed in the `ID` column have the structure `[parent_group, child_group, ..., key]`, and can be used to
index into the BaseBenchmarks suite to retrieve the corresponding benchmarks.
The percentages accompanying time and memory values in the below table are noise tolerances. The "true"
time/memory value for a given benchmark is expected to fall within this percentage of the reported value.
An empty cell means that the value was zero.

| ID                                                   | time            | GC time   | memory           | allocations |
|------------------------------------------------------|----------------:|----------:|-----------------:|------------:|
| `["composite", "kron(rand_const)"]`                  |    1.375 s (5%) |           |    2.73 KiB (1%) |          50 |
| `["composite", "kron(sparse_const)"]`                | 119.341 ms (5%) |           |   720 bytes (1%) |          15 |
| `["primitive", "S", "1"]`                            | 287.108 ns (5%) |           |   192 bytes (1%) |           4 |
| `["primitive", "S", "13"]`                           |  72.315 μs (5%) |           |    2.44 KiB (1%) |          52 |
| `["primitive", "S", "17"]`                           |   1.430 ms (5%) |           |    3.19 KiB (1%) |          68 |
| `["primitive", "S", "21"]`                           |  46.264 ms (5%) |           |    3.94 KiB (1%) |          84 |
| `["primitive", "S", "25"]`                           | 926.831 ms (5%) |           |    4.69 KiB (1%) |         100 |
| `["primitive", "S", "5"]`                            |   1.612 μs (5%) |           |   960 bytes (1%) |          20 |
| `["primitive", "S", "9"]`                            |   5.763 μs (5%) |           |    1.69 KiB (1%) |          36 |
| `["primitive", "Sdag", "1"]`                         | 288.084 ns (5%) |           |   192 bytes (1%) |           4 |
| `["primitive", "Sdag", "13"]`                        |  57.827 μs (5%) |           |    2.44 KiB (1%) |          52 |
| `["primitive", "Sdag", "17"]`                        |   1.079 ms (5%) |           |    3.19 KiB (1%) |          68 |
| `["primitive", "Sdag", "21"]`                        |  41.529 ms (5%) |           |    3.94 KiB (1%) |          84 |
| `["primitive", "Sdag", "25"]`                        | 882.400 ms (5%) |           |    4.69 KiB (1%) |         100 |
| `["primitive", "Sdag", "5"]`                         |   1.589 μs (5%) |           |   960 bytes (1%) |          20 |
| `["primitive", "Sdag", "9"]`                         |   4.938 μs (5%) |           |    1.69 KiB (1%) |          36 |
| `["primitive", "T", "1"]`                            | 286.028 ns (5%) |           |   192 bytes (1%) |           4 |
| `["primitive", "T", "13"]`                           |  49.932 μs (5%) |           |    2.44 KiB (1%) |          52 |
| `["primitive", "T", "17"]`                           | 954.401 μs (5%) |           |    3.19 KiB (1%) |          68 |
| `["primitive", "T", "21"]`                           |  38.874 ms (5%) |           |    3.94 KiB (1%) |          84 |
| `["primitive", "T", "25"]`                           | 782.382 ms (5%) |           |    4.69 KiB (1%) |         100 |
| `["primitive", "T", "5"]`                            |   1.581 μs (5%) |           |   960 bytes (1%) |          20 |
| `["primitive", "T", "9"]`                            |   4.800 μs (5%) |           |    1.69 KiB (1%) |          36 |
| `["primitive", "Tdag", "1"]`                         | 286.717 ns (5%) |           |   192 bytes (1%) |           4 |
| `["primitive", "Tdag", "13"]`                        |  58.703 μs (5%) |           |    2.44 KiB (1%) |          52 |
| `["primitive", "Tdag", "17"]`                        |   1.165 ms (5%) |           |    3.19 KiB (1%) |          68 |
| `["primitive", "Tdag", "21"]`                        |  41.392 ms (5%) |           |    3.94 KiB (1%) |          84 |
| `["primitive", "Tdag", "25"]`                        | 857.179 ms (5%) |           |    4.69 KiB (1%) |         100 |
| `["primitive", "Tdag", "5"]`                         |   1.590 μs (5%) |           |   960 bytes (1%) |          20 |
| `["primitive", "Tdag", "9"]`                         |   5.145 μs (5%) |           |    1.69 KiB (1%) |          36 |
| `["primitive", "TimeEvolution(heisenberg(n), 0.2)"]` |  998.271 s (5%) | 165.955 s | 1140.76 GiB (1%) |      128606 |
| `["primitive", "X", "1"]`                            | 279.000 ns (5%) |           |   192 bytes (1%) |           4 |
| `["primitive", "X", "13"]`                           |  89.471 μs (5%) |           |    2.44 KiB (1%) |          52 |
| `["primitive", "X", "17"]`                           |   1.777 ms (5%) |           |    3.19 KiB (1%) |          68 |
| `["primitive", "X", "21"]`                           |  65.499 ms (5%) |           |    3.94 KiB (1%) |          84 |
| `["primitive", "X", "25"]`                           |    1.414 s (5%) |           |    4.69 KiB (1%) |         100 |
| `["primitive", "X", "5"]`                            |   1.600 μs (5%) |           |   960 bytes (1%) |          20 |
| `["primitive", "X", "9"]`                            |   6.445 μs (5%) |           |    1.69 KiB (1%) |          36 |
| `["primitive", "Y", "1"]`                            | 277.555 ns (5%) |           |   192 bytes (1%) |           4 |
| `["primitive", "Y", "13"]`                           | 183.652 μs (5%) |           |    2.44 KiB (1%) |          52 |
| `["primitive", "Y", "17"]`                           |   3.776 ms (5%) |           |    3.19 KiB (1%) |          68 |
| `["primitive", "Y", "21"]`                           |  88.390 ms (5%) |           |    3.94 KiB (1%) |          84 |
| `["primitive", "Y", "25"]`                           |    1.955 s (5%) |           |    4.69 KiB (1%) |         100 |
| `["primitive", "Y", "5"]`                            |   1.798 μs (5%) |           |   960 bytes (1%) |          20 |
| `["primitive", "Y", "9"]`                            |  10.859 μs (5%) |           |    1.69 KiB (1%) |          36 |
| `["primitive", "Z", "1"]`                            | 284.374 ns (5%) |           |   192 bytes (1%) |           4 |
| `["primitive", "Z", "13"]`                           |  39.115 μs (5%) |           |    2.44 KiB (1%) |          52 |
| `["primitive", "Z", "17"]`                           | 754.080 μs (5%) |           |    3.19 KiB (1%) |          68 |
| `["primitive", "Z", "21"]`                           |  35.941 ms (5%) |           |    3.94 KiB (1%) |          84 |
| `["primitive", "Z", "25"]`                           | 719.002 ms (5%) |           |    4.69 KiB (1%) |         100 |
| `["primitive", "Z", "5"]`                            |   1.570 μs (5%) |           |   960 bytes (1%) |          20 |
| `["primitive", "Z", "9"]`                            |   4.262 μs (5%) |           |    1.69 KiB (1%) |          36 |

## Benchmark Group List
Here's a list of all the benchmark groups executed by this job:

- `["composite"]`
- `["primitive", "S"]`
- `["primitive", "Sdag"]`
- `["primitive", "T"]`
- `["primitive", "Tdag"]`
- `["primitive"]`
- `["primitive", "X"]`
- `["primitive", "Y"]`
- `["primitive", "Z"]`

## Julia versioninfo
```
Julia Version 1.1.0
Commit 80516ca202 (2019-01-21 21:24 UTC)
Platform Info:
  OS: macOS (x86_64-apple-darwin18.5.0)
  uname: Darwin 18.5.0 Darwin Kernel Version 18.5.0: Mon Mar 11 20:40:32 PDT 2019; root:xnu-4903.251.3~3/RELEASE_X86_64 x86_64 i386
  CPU: Intel(R) Core(TM) i7-7700HQ CPU @ 2.80GHz: 
              speed         user         nice          sys         idle          irq
       #1  2800 MHz     419096 s          0 s     247485 s    3506710 s          0 s
       #2  2800 MHz      41047 s          0 s      26379 s    4105423 s          0 s
       #3  2800 MHz     315433 s          0 s     153254 s    3704164 s          0 s
       #4  2800 MHz      39299 s          0 s      19046 s    4114504 s          0 s
       #5  2800 MHz     273904 s          0 s     115510 s    3783436 s          0 s
       #6  2800 MHz      39975 s          0 s      17396 s    4115476 s          0 s
       #7  2800 MHz     255755 s          0 s      91304 s    3825790 s          0 s
       #8  2800 MHz      42279 s          0 s      16566 s    4114001 s          0 s
       
  Memory: 16.0 GB (8600.39453125 MB free)
  Uptime: 1.024235e6 sec
  Load Avg:  2.255859375  2.19873046875  2.1748046875
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-6.0.1 (ORCJIT, skylake)
```