@GiggleLiu 2019-05-20
Cool, could you please also open a pr for the old 0.3.2 branch to show the performance difference?