@Roger-luo 2019-05-20
Let me fix this performance issue first. So the slow down is mainly because we start to use `apply!` and treat Hamiltonian as blocks with `sum`, which is quite inefficient since `sum` has to copy the register each time it is applied.

Now I changed the inner structure of `TimeEvolution` to let it take the matrix directly, the performance improves to be 4x faster than what we had in `v0.3.2`, that's probably why @GiggleLiu dropped the `LinearMap` in `v0.3.2`.

Before:

```
julia> te = TimeEvolution(heisenberg(10), 0.2);

julia> @benchmark apply!($(rand_state(10)), $te)
BenchmarkTools.Trial:
  memory estimate:  6.87 MiB
  allocs estimate:  863
  --------------
  minimum time:     2.903 ms (0.00% GC)
  median time:      3.531 ms (0.00% GC)
  mean time:        4.145 ms (17.74% GC)
  maximum time:     71.263 ms (94.55% GC)
  --------------
  samples:          1205
  evals/sample:     1
```

After:

```
julia> @benchmark apply!($r, $te)
BenchmarkTools.Trial: 
  memory estimate:  560.53 KiB
  allocs estimate:  167
  --------------
  minimum time:     572.927 μs (0.00% GC)
  median time:      861.727 μs (0.00% GC)
  mean time:        867.940 μs (6.87% GC)
  maximum time:     58.869 ms (98.41% GC)
  --------------
  samples:          5742
  evals/sample:     1
```

And I think since Hamiltonian is not the actually circuit, using the circuit as its representation is kinda ill defined, and not necessary. This new structure make things much simpler and will allow user defined Hamiltonian from other package to make use of Yao in a simpler way.