@Roger-luo 2019-05-20
And actually I don't see much use cases to have a much more general `LinearMap` for Hamiltonians, thus I believe simper is better.