@Roger-luo 2019-05-21
BTW, I fixed a small bug for doing circuit simplification.